feat: add source map implementation

+++ cyc/src/parse/lexer/cursor/cursor_impl.rs
@@ -61,8 +61,8 @@ impl<'src> Cursor<'src> {
     }
 
     /// Returns amount of already bumped symbols.
-    pub(super) fn bumped_len(&self) -> usize {
-        self.len_remaining - self.chars.as_str().len()
+    pub(super) fn bumped_len(&self) -> u32 {
+        (self.len_remaining - self.chars.as_str().len()) as u32
     }
 
     /// Resets the number of bytes consumed to 0.
